Add some feature to RPS * fix some layout in RPS * now can show how many revision were made and when they where made * there is still a bug in the correlation table section

+++ database/migrations/2025_07_31_232926_create_rps_topics_table.php
@@ -37,7 +37,7 @@ return new class extends Migration
             $table->text('materi_pembelajaran')->nullable();
 
             // Kolom "Bobot Penilaian (%)"
-            $table->integer('bobot_penilaian')->default(0);
+            $table->float('bobot_penilaian')->default(0);
 
             // Timestamps standar Laravel
             $table->timestamps();
